pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "poultry-caduceus"
version = "1.0.0"
description = "A Bidirectional DNA Language Model for Chicken Genome"
readme = "README.md"
license = {text = "MIT"}
authors = [
    {name = "Your Name", email = "your.email@example.com"}
]
maintainers = [
    {name = "Your Name", email = "your.email@example.com"}
]
keywords = [
    "deep learning",
    "genomics", 
    "DNA language model",
    "chicken genome",
    "poultry",
    "bioinformatics",
    "MPRA",
    "regulatory genomics"
]
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Science/Research",
    "License :: OSI Approved :: MIT License",
    "Operating System :: OS Independent",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Topic :: Scientific/Engineering :: Artificial Intelligence",
    "Topic :: Scientific/Engineering :: Bio-Informatics",
]
requires-python = ">=3.9"
dependencies = [
    "torch>=2.0.0",
    "numpy>=1.24.0",
    "pandas>=2.0.0",
    "scipy>=1.10.0",
    "scikit-learn>=1.2.0",
    "transformers>=4.30.0",
    "einops>=0.6.0",
    "tqdm>=4.65.0",
    "pyyaml>=6.0",
    "biopython>=1.81",
    "pyarrow>=12.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0",
    "pytest-cov>=4.0",
    "black>=23.0",
    "isort>=5.12",
    "flake8>=6.0",
    "mypy>=1.0",
    "pre-commit>=3.0",
]
docs = [
    "sphinx>=6.0",
    "sphinx-rtd-theme>=1.2",
    "myst-parser>=1.0",
    "sphinx-autodoc-typehints>=1.23",
]
training = [
    "wandb>=0.15.0",
    "tensorboard>=2.13.0",
    "accelerate>=0.20.0",
    "deepspeed>=0.9.0",
]
all = [
    "poultry-caduceus[dev,docs,training]",
]

[project.scripts]
poultry-caduceus = "poultry_caduceus.cli:main"

[project.urls]
Homepage = "https://github.com/YOUR_USERNAME/PoultryCaduceus"
Documentation = "https://github.com/YOUR_USERNAME/PoultryCaduceus#documentation"
Repository = "https://github.com/YOUR_USERNAME/PoultryCaduceus"
Issues = "https://github.com/YOUR_USERNAME/PoultryCaduceus/issues"

[tool.setuptools.packages.find]
where = ["."]
include = ["poultry_caduceus*"]
exclude = ["tests*", "examples*", "scripts*", "docs*"]

[tool.setuptools.package-data]
poultry_caduceus = ["configs/*.yaml"]

[tool.black]
line-length = 100
target-version = ['py39', 'py310', 'py311']
include = '\.pyi?$'
exclude = '''
/(
    \.eggs
  | \.git
  | \.hg
  | \.mypy_cache
  | \.tox
  | \.venv
  | _build
  | buck-out
  | build
  | dist
)/
'''

[tool.isort]
profile = "black"
line_length = 100
known_first_party = ["poultry_caduceus"]
skip = [".git", ".venv", "venv", "build", "dist"]

[tool.mypy]
python_version = "3.9"
warn_return_any = true
warn_unused_configs = true
ignore_missing_imports = true
exclude = ["tests", "docs", "examples"]

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
python_functions = ["test_*"]
addopts = "-v --cov=poultry_caduceus --cov-report=term-missing"
filterwarnings = [
    "ignore::DeprecationWarning",
    "ignore::UserWarning",
]

[tool.coverage.run]
source = ["poultry_caduceus"]
omit = ["tests/*", "examples/*", "scripts/*"]

[tool.coverage.report]
exclude_lines = [
    "pragma: no cover",
    "def __repr__",
    "raise AssertionError",
    "raise NotImplementedError",
    "if __name__ == .__main__.:",
]
